Added ICgen 'glass' position gen method based on James Wadsley's method
For an IC object the user can now select IC.settings.pos_gen.method='glass'.
This is now the preferred and default method as it generates less noisy ICs.

From the doc string on the method:

Generates x-y positions and z-values mapped to -1 to 1 by tiling a
small peridic cube glass and mapping to physical space.

This is done by generating a uniformly distributed glass cylinder in
what I'll call CDF space, for lack of a better word.  By using the
inverse radial and vertical CDFs we can map points on that cylinder to
points in real space which will be distributed according to the PDF
(i.e. density) in real space.

Since the ICs are vaguely cylindrical, this can be done without
distorting the glass too-much (locally).  The final result is
significantly more glass-like than using other methods.

The only trick is to approximate the height of the cylinder in CDF
space in a reasonable way.  One method that works is to just
take the average <H/R>, mass weighted over the disk (For a unit radius
cylinder).

I use a slightly more convoluted method which seems to work pretty well,
but either way should be good.